works on vc7++, cuurently an exe target, must point project to openssl.  Copy config.h_win32vc7 to config.h

+++ crypto/math/stat.c
@@ -208,7 +208,7 @@ stat_test_rand_source(rand_source_func_t get_rand_bytes) {
   uint16_t gaps[6] = { 0, 0, 0, 0, 0, 0 };
   uint16_t lo_value[6] = { 2315, 1114, 527, 240, 103, 103 };
   uint16_t hi_value[6] = { 2685, 1386, 723, 384, 209, 209 };
-  int16_t  state = 0;
+  int state = 0;
   uint16_t mask;
   
   /* counters for monobit, poker, and runs tests are initialized above */
